initial commit
Status: Working. This script reads the boiler temperature, converts that value to an RGB value, 
sends it to the LEDs and then sends the temp over serial to the Pi
TODO Need to add comments, clean up old pattern references and any other left over code from base file
This code is just a hacked version of the FastLED DemoReel100 by Mark Kriegsman.